remove getAccountAndProfile in favour of profile.decrypt
