+++ src/commands/transaction/accountlink.ts
@@ -48,7 +48,8 @@ export default class extends AnnounceTransactionsCommand {
     }
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         options.publicKey = OptionsResolver(options,
             'publicKey',
