+++ src/commands/transaction/accountmosaicrestriction.ts
@@ -60,7 +60,8 @@ export default class extends AnnounceTransactionsCommand {
 
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         options.restrictionFlag = OptionsResolver(options,
             'restrictionFlag',
