+++ src/commands/transaction/addressalias.ts
@@ -55,7 +55,8 @@ export default class extends AnnounceTransactionsCommand {
 
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         options.namespace = OptionsResolver(options,
             'namespace',
