+++ src/commands/transaction/mosaic.ts
@@ -95,7 +95,8 @@ export default class extends AnnounceTransactionsCommand {
 
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         const nonce = MosaicNonce.createRandom();
         let blocksDuration;
