+++ src/commands/transaction/mosaicsupplychange.ts
@@ -57,7 +57,8 @@ export default class extends AnnounceTransactionsCommand {
 
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         options.mosaicId = OptionsResolver(options,
             'mosaicId',
