+++ src/commands/transaction/namespace.ts
@@ -67,7 +67,8 @@ export default class extends AnnounceTransactionsCommand {
 
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         options.name = OptionsResolver(options,
             'name',
