+++ src/commands/transaction/transfer.ts
@@ -92,7 +92,8 @@ export default class extends AnnounceTransactionsCommand {
 
     @metadata
     execute(options: CommandOptions) {
-        const { account, profile } = this.getAccountAndProfile(options);
+        const profile = this.getProfile(options);
+        const account = profile.decrypt(options);
 
         const recipientAddress = AccountService.getRecipient(OptionsResolver(options,
             'recipientAddress',
